Move GetMeshTextureAssets into AssetLookup
